Station changes due to changing tick counter to bigger size

+++ src/base_station_base.h
@@ -52,7 +52,7 @@ struct StationRect : public Rect {
 struct BaseStation : StationPool::PoolItem<&_station_pool> {
 	TileIndex xy;                   ///< Base tile of the station
 	TrackedViewportSign sign;       ///< NOSAVE: Dimensions of sign
-	byte delete_ctr;                ///< Delete counter. If greater than 0 then it is decremented until it reaches 0; the waypoint is then is deleted.
+	int delete_ctr;                ///< Delete counter. If greater than 0 then it is decremented until it reaches 0; the waypoint is then is deleted.
 
 	std::string name;               ///< Custom name
 	StringID string_id;             ///< Default name (town area) of station
